app/scripts/requirements-whisper.txt
openai-whisper
flask
flask-cors
numpy
torch
deep-translator
pyannote.audio
soundfile
